requirements.txt
pytorch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
pytorch-cuda=12.4
e3nn
torch-geometric
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv
pandas
numpy
scipy
scikit-learn
tqdm
mendeleev
nltk
openai
rouge-score
gotennet
egnn-pytorch
